requirements.txt
streamlit>=1.25.0
python-dotenv>=1.0.0
groq>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
